Rename configfile -> config/file
