+++ config/file/raw.go
@@ -1,4 +1,4 @@
-package configfile
+package file
 
 // rawConfig is a raw data model for runner config files.
 // It serves as a receiver for unmarshaling processes and for that reason
